removed additional folder in structure
