+++ frontend/src/App.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function App() {
+  return (
+      <h1>Hello World</h1>
+  );
+}
+
+export default App;
